requirements.txt
azure-identity
azure-keyvault-certificates
azure-keyvault-secrets
certbot
azure-storage-blob
cryptography
